Extract region pipeline coordinators
